Move repeated code into utility function
