+++ lib/annotations/sizes.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const css2obj = require('../utils/css2obj.js');
+
 /**
  * Custom `@sizes` annotation. Expects sizes type followed by
  * semicolon-separated list of `key: value` arguments.
@@ -17,15 +19,6 @@ module.exports = () => ({
     const options = {
       type: 'sizes',
     };
-    let key, value;
-    raw.split(';').forEach(option => {
-      const parts = option.split(':');
-      key = parts[0].trim();
-      value = parts[1] ? parts[1].trim() : null;
-      if (key) {
-        options[key] = value;
-      }
-    });
-    return options;
+    return css2obj(raw, options);
   },
 });
